feat : order 관련 insert, select 및 junit 테스트 개발

+++ orderShop/src/main/java/com/cjy9249/orderShop/controller/user/UserController.java
@@ -18,7 +18,7 @@ public class UserController {
 
     @GetMapping("info")
     public ResponseEntity<Response<User>> getUserInfo(
-            @RequestParam(value = "userId") int userId
+            @RequestParam(value = "userId") Long userId
     ) {
         return ResponseEntity.ok(
                 Response.of(
