+++ orderShop/src/main/java/com/cjy9249/orderShop/domain/repository/order/OrderRepository.java
@@ -4,5 +4,5 @@ import com.cjy9249.orderShop.domain.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order,Integer>, OrderRepositoryCustom {
-    Order findByOrderId(int orderId);
+    Order findByOrderId(Long orderId);
 }
