+++ orderShop/src/main/java/com/cjy9249/orderShop/domain/repository/user/UserRepository.java
@@ -4,5 +4,5 @@ import com.cjy9249.orderShop.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Integer>, UserRepositoryCustom{
-    User findByUserId(int userId);
+    User findByUserId(Long userId);
 }
