+++ orderShop/src/main/java/com/cjy9249/orderShop/service/user/GetUserInfoService.java
@@ -11,7 +11,7 @@ public class GetUserInfoService {
     private final UserRepository userRepository;
 
     @Transactional(readOnly = true)
-    public User getUserInfo(int userId) {
+    public User getUserInfo(Long userId) {
         return userRepository.findByUserId(userId);
     }
 }
